refactor: split agent-hook-executor into focused modules (960→6 files)
Extract 5 new modules from the 960-line monolith:
- pr-verification.js: GitHub PR verification with retry logic
- hook-sandbox.js: Shared VM sandbox builder (deduplicated)
- hook-transform.js: Transform script execution
- hook-template.js: Template variable substitution
- hook-logic.js: Hook logic evaluation

Main file becomes slim dispatcher (116 lines) importing from modules.
All files pass strict ESLint (300 max lines, complexity 10, depth 3).

Key bug fix: PR fetch now retries 6 times (5s apart) for concrete PR
numbers, preventing false HALLUCINATED errors caused by GitHub API
eventual consistency after merge + branch delete.
